backend/requirements.txt
asgiref==3.11.1
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2026.1.post1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
